Fix bugprone narrowing conversions

+++ cpr/curlholder.cpp
@@ -29,7 +29,7 @@ CurlHolder::~CurlHolder() {
 
 std::string CurlHolder::urlEncode(const std::string& s) const {
     assert(handle);
-    char* output = curl_easy_escape(handle, s.c_str(), s.length());
+    char* output = curl_easy_escape(handle, s.c_str(), static_cast<int>(s.length()));
     if (output) {
         std::string result = output;
         curl_free(output);
@@ -40,7 +40,7 @@ std::string CurlHolder::urlEncode(const std::string& s) const {
 
 std::string CurlHolder::urlDecode(const std::string& s) const {
     assert(handle);
-    char* output = curl_easy_unescape(handle, s.c_str(), s.length(), nullptr);
+    char* output = curl_easy_unescape(handle, s.c_str(), static_cast<int>(s.length()), nullptr);
     if (output) {
         std::string result = output;
         curl_free(output);
